# packages/fixit-ai/fixit_ai-0.0.1.tar.gz/fixit_ai-0.0.1/examples/express_buggy/.fixit/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Express API",
    "version": "1.0.0",
    "description": "Auto-generated OpenAPI spec from Express.js routes"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/users": {
      "post": {
        "summary": "POST /api/users",
        "description": "Auto-generated from Express route: POST /api/users",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "POST /api/auth/login",
        "description": "Auto-generated from Express route: POST /api/auth/login",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/users/:id": {},
    "/api/users/{id}": {
      "get": {
        "summary": "GET /api/users/:id",
        "description": "Auto-generated from Express route: GET /api/users/:id",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: id"
          }
        ]
      }
    },
    "/health": {
      "get": {
        "summary": "GET /health",
        "description": "Auto-generated from Express route: GET /health",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  }
}